requirements.txt
discord.py
PyNaCl
python-dotenv
lyricsgenius
yt-dlp[default]
